Move WebView-related stuff into Foreign.JavaScript.TH
